WIP: saving DeviceMap with Snapshot, including new 'missing' state
First attempt at refactoring the DeviceMap and project management in
order to save the state of the DeviceMap. Sova will try to remember
about all the virtual / physical MIDI slots and OSC slots used and
restore them when loading a project. Needs to be refactored to minimize
changes to core and to make the logic as simple as possible.